pyproject.toml
[tool.poetry]
name = "wenling"
version = "0.1.0"
description = ""
authors = ["Yx Jiang <2237303+yxjiang@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
openai = "^1.3.6"
fastapi = "^0.104.1"
notion-client = "^2.1.0"
retrying = "^1.3.4"
pytz = "^2023.3.post1"
colorama = "^0.4.6"
requests = "^2.31.0"
types-colorama = "^0.4.15.12"
pydantic = "^2.5.2"
bs4 = "^0.0.1"
aiohttp = "^3.9.1"
types-requests = "^2.31.0.10"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120